==> 3_evaluation/1_ground_truth_data.ipynb <==
Get for each original question 5 similar questions for evaluation

import requests 

docs_url = 'https://github.com/DataTalksClub/llm-zoomcamp/blob/main/01-intro/documents.json?raw=1'
docs_response = requests.get(docs_url)
documents_raw = docs_response.json()

documents = []

for course in documents_raw:
    course_name = course['course']

    for doc in course['documents']:
        doc['course'] = course_name
        documents.append(doc)

import hashlib

def generate_document_id(doc):
    # combined = f"{doc['course']}-{doc['question']}"
    # for original hash
    combined = f"{doc['course']}-{doc['question']}-{doc['text'][:10]}"
    hash_object = hashlib.md5(combined.encode())
    hash_hex = hash_object.hexdigest()
    document_id = hash_hex[:8]
    return document_id

for doc in documents:
    doc['id'] = generate_document_id(doc)

documents[3]

from collections import defaultdict

hashes = defaultdict(list)

for doc in documents:
    doc_id = doc['id']
    hashes[doc_id].append(doc)

len(hashes), len(documents)

# we have only 1 collision
for k, values in hashes.items():
    if len(values) > 1:
        print(k, len(values))

hashes['593f7569']

import json

with open('documents-with-ids.json', 'wt') as f_out:
    json.dump(documents, f_out, indent=2)

!head documents-with-ids.json

prompt_template = """
You emulate a student who's taking our course.
Formulate 5 questions this student might ask based on a FAQ record. The record
should contain the answer to the questions, and the questions should be complete and not too short.
If possible, use as fewer words as possible from the record. 

The record:

section: {section}
question: {question}
answer: {text}

Provide the output in parsable JSON without using code blocks:

["question1", "question2", ..., "question5"]
""".strip()

from openai import OpenAI
client = OpenAI()

def generate_questions(doc):
    prompt = prompt_template.format(**doc)

    response = client.chat.completions.create(
        model='gpt-4o',
        messages=[{"role": "user", "content": prompt}]
    )

    json_response = response.choices[0].message.content
    return json_response

from tqdm.auto import tqdm

results = {}

for doc in tqdm(documents): 
    doc_id = doc['id']
    # if already was generated for this id
    if doc_id in results:
        continue

    questions = generate_questions(doc)
    results[doc_id] = questions

import pickle

with open('results.bin', 'rb') as f_in:
    results = pickle.load(f_in)

results['1f6520ca']

parsed_resulst = {}

for doc_id, json_questions in results.items():
    parsed_resulst[doc_id] = json.loads(json_questions)

doc_index = {d['id']: d for d in documents}

final_results = []

for doc_id, questions in parsed_resulst.items():
    course = doc_index[doc_id]['course']
    for q in questions:
        final_results.append((q, course, doc_id))

import pandas as pd

df = pd.DataFrame(final_results, columns=['question', 'course', 'document'])

df.to_csv('ground-truth-data.csv', index=False)

!head ground-truth-data.csv

We have question, for what course it is, and id of document
